favicon, logo 이미지 추가, App.js, google loginapi 구조 변경, 로그인 메인페이지 구현
