+++ src/containers/style/main.css
@@ -1,0 +1,6 @@
+.header{
+    font-size:22px;
+    color:teal;
+    text-align: center;
+    font-style: italic;
+}
