+++ src/index.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App'
-
+import App from './components/App'
 import './index.css';
-//import App from './App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render( <App/>, document.getElementById('root'));
